pyproject.toml
[project]
name = "robosystems"
version = "1.2.5"
description = "RoboSystems"
requires-python = ">=3.13"
authors = [
    {name = "RFS LLC"},
]
license = {text = "MIT"}
keywords = ["robosystems", "financial", "graph", "knowledge graph", "ladybugdb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Topic :: Office/Business :: Financial :: Accounting",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Database",
    "Topic :: Database :: Database Engines/Servers",
    "Typing :: Typed",
]
dependencies = [
    # Web Framework & API
    "fastapi>=0.116.0,<1.0",
    "fastapi-limiter>=0.1.6,<1.0",
    "python-multipart>=0.0.16,<1.0",
    "sse-starlette>=2.4.1,<3.0",
    "uvicorn>=0.35.0,<1.0",

    # Graph Databases
    "real-ladybug==0.12.2",
    "neo4j>=5.15.0,<6.0",
    "duckdb>=1.1.0,<2.0",

    # Relational Database (PostgreSQL)
    "alembic>=1.16.0,<2.0",
    "psycopg2-binary>=2.9.0,<3.0",
    "sqlalchemy>=2.0.0,<3.0",

    # Caching
    "redis>=6.2.0,<7.0",

    # Data Orchestration (Dagster)
    "dagster>=1.9.0,<2.0",
    "dagster-webserver>=1.9.0,<2.0",
    "dagster-postgres>=0.25.0,<1.0",
    "dagster-aws>=0.25.0,<1.0",

    # Cloud Services (AWS)
    "boto3>=1.39.0,<2.0",

    # Data Processing & Analytics
    "arelle-release==2.37.12",
    "pandas>=2.3.0,<3.0",
    "pyarrow>=20.0.0",

    # External API Integrations
    "intuit-oauth>=1.2.0,<2.0",
    "plaid-python>=34.0,<35.0",
    "python-quickbooks>=0.9.0,<1.0",
    "stripe>=11.1.0,<12.0",

    # HTTP & Async Clients
    "aiohttp>=3.12.0",
    "async_timeout>=5.0.0,<6.0",
    "httpx>=0.28.0",
    "httpx-sse>=0.4.0",
    "requests>=2.32.0,<3.0",

    # Observability & Monitoring
    "opentelemetry-api>=1.35.0,<2.0",
    "opentelemetry-exporter-otlp>=1.35.0,<2.0",
    "opentelemetry-instrumentation-aiohttp-client>=0.55b0,<1.0",
    "opentelemetry-instrumentation-fastapi>=0.55b0,<1.0",
    "opentelemetry-instrumentation-psycopg2>=0.55b0,<1.0",
    "opentelemetry-instrumentation-requests>=0.55b0,<1.0",
    "opentelemetry-sdk>=1.35.0,<2.0",

    # Security & Authentication
    "bcrypt>=4.3.0,<5.0",
    "email-validator>=2.2.0,<3.0",
    "pyjwt>=2.10.0,<3.0",

    # Data Serialization & Identifiers
    "pydantic[email]>=2.11.0,<3.0",
    "python-ulid>=3.0.0",
    "pyyaml>=6.0.0,<7.0",
    "uuid6>=2025.0.0",

    # XML/HTML Processing
    "beautifulsoup4>=4.13.0,<5.0",
    "lxml>=5.4.0,<6.0",

    # Utilities
    "holidays>=0.76.0,<1.0",
    "psutil>=6.1.0,<7.0",
    "python-slugify>=8.0.0,<9.0",
    "retrying>=1.4.0,<2.0",
]

[project.optional-dependencies]
dev = [
    # RoboSystems client for demo and testing
    "robosystems-client==0.2.22",

    # Testing framework
    "pytest>=8.4.0,<9.0",
    "pytest-asyncio>=1.0.0,<2.0",
    "pytest-cov>=6.2.0",
    "pytest-env>=1.1.0,<2.0",
    "pytest-mock>=3.14.0,<4.0",

    # Code quality tools
    "ruff>=0.12.0,<1.0",
    "basedpyright>=1.22.0,<2.0",

    # AWS development tools
    "awscli-local>=0.22.0,<1.0",
    "cfn-lint>=1.37.0,<2.0",
    "moto[s3]>=5.1.6,<6.0",

    # Development utilities
    "jupyter>=1.1.0,<2.0",
    "rich>=14.0.0,<15.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["robosystems"]

[tool.setuptools.package-data]
robosystems = ["**/*.py"]

[tool.setuptools.exclude-package-data]
"*" = ["data/*", "notebooks/*", "cloudformation/*"]

[tool.ruff]
exclude = [
    ".git",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "examples",
    "robosystems/adapters/sec/arelle",
]
line-length = 88
indent-width = 2
target-version = "py313"

[tool.ruff.lint]
select = [
    "E4", "E7", "E9", "F",  # Original: pyflakes, pycodestyle errors
    "I",      # isort - import sorting
    "B",      # bugbear - common bugs and design problems
    "C4",     # flake8-comprehensions - better comprehensions
    "UP",     # pyupgrade - modern Python syntax
    "S",      # bandit - security issues
    "SIM",    # flake8-simplify - code simplification
    "RUF",    # ruff-specific rules
]
ignore = [
    # Security - acceptable patterns for this codebase
    "S101",   # assert usage (needed for tests)
    "S104",   # binding to all interfaces (intentional in containers)
    "S105",   # hardcoded password detection (false positives)
    "S106",   # hardcoded password in func args (false positives)
    "S107",   # hardcoded password default (false positives in tests)
    "S110",   # try-except-pass (often intentional)
    "S112",   # try-except-continue (intentional in loaders)
    "S202",   # tarfile.extractall (controlled input - Arelle cache)
    "S310",   # suspicious URL open (file: scheme intentional for Arelle)
    "S311",   # pseudo-random generators (acceptable for non-crypto)
    "S324",   # insecure hash MD5 (used for cache keys, not security)
    "S602",   # subprocess shell=True (internal admin CLI only)
    "S603",   # subprocess call (internal tools, not user input)
    "S607",   # partial executable path (internal tools)
    "S608",   # hardcoded-sql-expression (Cypher queries, not SQL)
    "S108",   # hardcoded-temp-file (common in tests)
    # Bugbear - patterns acceptable in this codebase
    "B007",   # unused-loop-control-variable (style)
    "B008",   # function call in default argument (FastAPI Depends pattern)
    "B017",   # assert-raises-exception (test pattern)
    "B023",   # function-uses-loop-variable (usually fine)
    "B904",   # raise-without-from-inside-except (nice-to-have, not a bug)
    # Simplify - style preferences
    "SIM102", # collapsible-if (style preference)
    "SIM103", # needless-bool (style preference)
    "SIM105", # suppressible-exception (style preference)
    "SIM108", # if-else-block-instead-of-if-exp (style preference)
    "SIM113", # enumerate-for-loop (style preference)
    "SIM116", # if-else-block-instead-of-dict (style preference)
    "SIM117", # multiple-with-statements (often clearer as separate)
    # Ruff-specific - patterns acceptable in this codebase
    "RUF001", # ambiguous-unicode-character (intentional in tests)
    "RUF006", # asyncio-dangling-task (fire-and-forget pattern)
    "RUF012", # mutable-class-default (config classes, immutable in practice)
    "RUF034", # useless-if-else (minor)
    "RUF059", # unused unpacked variable (common tuple unpacking pattern)
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.basedpyright]
include = ["robosystems", "main.py"]
exclude = ["**/__pycache__", ".venv", "examples", "robosystems/adapters/sec/arelle", "tests"]
extraPaths = ["."]
pythonVersion = "3.13"
venvPath = "."
venv = ".venv"
typeCheckingMode = "standard"
reportArgumentType = "none"
reportAttributeAccessIssue = "none"
reportGeneralTypeIssues = "none"
reportOptionalMemberAccess = "none"
reportReturnType = "none"
reportCallIssue = "none"
reportMissingImports = "none"
